==> notebooks/TIRE_brain_human/2A_Clustering.Rmd <==
---
title: "Human neurosphere sample clustering"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      out.width="150%",
                      fig.path='./figures/clustering/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())

library(SingleCellExperiment)
library(tidyverse)
library(scuttle)
library(scran)
library(scater)
library(patchwork)
library(knitr)
library(here)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# TIRE-Seq human brain neurosphere clustering

## Recap

These are the first experiments where I use version 2 chemistry. Read 2 is now the barcode and UMI. Read 1 is the cDNA. This was done to avoid reading through the TSO on Ilumina read 1

This sequencing run consists of 3 experiments:

1. Test the version 2 protocol with magnetic poly-T Dynabeads [https://au-mynotebook.labarchives.com/share/Piper_Research_Project/MjAuOHwxNzE3NjUvMTYvVHJlZU5vZGUvNDE3OTUxMTA4N3w1Mi44](ELN link])
2. Benchmark the version 2 protocol with a mixture of human and mouse cell lines Daniel Brown generated [https://au-mynotebook.labarchives.com/share/Piper_Research_Project/MjkuOTAwMDAwMDAwMDAwMDAyfDE3MTc2NS8yMy9UcmVlTm9kZS8yMzM4NDg4MTE1fDc1Ljg5OTk5OTk5OTk5OTk5](ELN link)
3. Test version 2 protocol with a biological appliation from Zac Moore of Brain Cancer lab [https://au-mynotebook.labarchives.com/share/Piper_Research_Project/MzEuMjAwMDAwMDAwMDAwMDAzfDE3MTc2NS8yNC9UcmVlTm9kZS8zMzczNDM0NjE1fDc5LjE5OTk5OTk5OTk5OTk5](ELN link)

Check the samples cluster by their cell type

## Read SCE and preprocessing

This was generated in notebook 1A_generateSCE.

```{r}
sce <- readRDS(here::here(
  "data/TIRE_brain_human/SCEs/", "brainCancer_basic.sce.rds"
   ))
```

Library size normalization and transformation

```{r, eval=T}
set.seed(666)

lib.sf <- librarySizeFactors(sce)
sce <- computeSumFactors(sce)
sce <- logNormCounts(sce)

```

### Feature selection

Here we follow the example of https://bioconductor.org/books/release/OSCA/feature-selection.html#variance-of-the-log-counts


```{r eval=T}
dec.sce <- modelGeneVar(sce)
fit.sce <- metadata(dec.sce)
hvg.sce.var <- getTopHVGs(dec.sce, n=1000)
sce <- scater::runPCA(sce, subset_row=hvg.sce.var, ncomponents=10)
```

Visulaise the fit. According to OSCA book:

*At any given abundance, we assume that the variation in expression for most genes is driven by uninteresting processes like sampling noise. Under this assumption, the fitted value of the trend at any given gene’s abundance represents an estimate of its uninteresting variation, which we call the technical component.*

Therefore the efitted line represents technical variation of which is much higher in version 1 where I used the **read object** compared to version 2 where the **UMI** object is used.

*We then define the biological component for each gene as the difference between its total variance and the technical component. This biological component represents the “interesting” variation for each gene and can be used as the metric for HVG selection.*

Build the object

```{r}
tb <- as_tibble(cbind(fit.sce$mean, fit.sce$var))

colnames(tb) <- c("Mean", "Variance")
```

Make the plot of technical and biological variation.

```{r tech_variation, fig.cap="The coloured line represents technical variation"}
plt1 <- ggplot(tb, 
             aes(x = Mean, y= Variance)) + 
  geom_point(alpha = 0.2, size=0.5) + 
  guides(colour = guide_legend(override.aes = list(size=2, alpha=1))) +
  xlab("Mean of log-expression") + 
  ylab("Variance of log-expression") +
  scale_colour_brewer(type="qualitative", palette = "Dark2") +
  geom_function(fun = fit.sce$trend, colour = "darkgreen") +
  theme_Publication(base_size = 16)

plt1
```

## PCA plots all samples 

This includes my samples too. Separation is very clearly human and mouse.

```{r PCA_all_samples}
plt1 <- plotPCA(sce, colour_by="Cell_Line") + theme_Publication()

plt1
```

## Filter for Brain cancer and remove low quality samples

This distort the PCA and should be removed anyway when I do DE

```{r lib_gene}
tb <- as_tibble(colData(sce))

plt1 <- ggplot(tb, 
             aes(x = sum, y= detected, colour=Storage)) + 
  geom_point(size=1.5) + 
  xlab("UMI counts") + 
  ylab("Genes detected") +
  scale_x_continuous(trans='log10') +
  annotation_logticks(base = 10, sides = "b") +
  ylim(0,20000) +
  scale_colour_brewer(type="qualitative", palette = "Dark2") +
  theme_Publication(base_size = 16)
```

Remove the low quality samples. Focus on the brain cancer samples that were stored frozen.

```{r}
sce <- sce[,sce$Researcher == "ZM"]
sce <- sce[,sce$Storage == "Freezer"]
sce <- sce[,sce$sum > 60000]
```

### Plot library size again after removing low quality samples.

```{r lib_gene_rm, fig.asp=1}
tb <- as_tibble(colData(sce))

plt2 <- ggplot(tb, 
             aes(x = sum, y= detected, colour=Storage)) + 
  geom_point(size=1.5) + 
  xlab("UMI counts") + 
  ylab("Genes detected") +
  scale_x_continuous(trans='log10') +
  annotation_logticks(base = 10, sides = "b") +
  ylim(0,20000) +
  scale_colour_brewer(type="qualitative", palette = "Dark2") +
  theme_Publication(base_size = 16)

plt1 / plt2
```

Library size normalization and transformation

```{r, eval=T}
set.seed(666)
lib.sf <- librarySizeFactors(sce)
sce <- computeSumFactors(sce)
sce <- logNormCounts(sce)
```

Feature selection

There is a lot more variation than the example at https://bioconductor.org/books/release/OSCA/feature-selection.html#variance-of-the-log-counts

Choose 1000 genes instead of standard 1000 because of the presence of human and mouse genes.

```{r eval=T}
dec.sce <- modelGeneVar(sce)
fit.sce <- metadata(dec.sce)
hvg.sce.var <- getTopHVGs(dec.sce, n=1000)
sce <- runPCA(sce, subset_row=hvg.sce.var, ncomponents=10)
```

Visulaise the fit. According to OSCA book:

*At any given abundance, we assume that the variation in expression for most genes is driven by uninteresting processes like sampling noise. Under this assumption, the fitted value of the trend at any given gene’s abundance represents an estimate of its uninteresting variation, which we call the technical component.*

Therefore the efitted line represents technical variation of which is much higher in version 1 where I used the **read object** compared to version 2 where the **UMI** object is used.

*We then define the biological component for each gene as the difference between its total variance and the technical component. This biological component represents the “interesting” variation for each gene and can be used as the metric for HVG selection.*

Build the object

```{r}
tb <- as_tibble(cbind(fit.sce$mean, fit.sce$var))
colnames(tb) <- c("Mean", "Variance")
```

### Plot mean variance of high quality glioma samples

```{r tech_variation_ZM_rm, fig.cap="The coloured line represents technical variation"}
plt1 <- ggplot(tb, 
             aes(x = Mean, y= Variance)) + 
  geom_point(alpha = 0.2, size=0.5) + 
  guides(colour = guide_legend(override.aes = list(size=2, alpha=1))) +
  xlab("Mean of log-expression") + 
  ylab("Variance of log-expression") +
  scale_colour_brewer(type="qualitative", palette = "Dark2") +
  geom_function(fun = fit.sce$trend, colour = "darkgreen") +
  theme_Publication(base_size = 16)

plt1
```

## PCA plots high quality glioma samples

Convert some drug metadata to factors

```{r}
sce$Dose_M <- as.numeric(sce$Dose_M)
sce$Day_Exposure <- as.factor(sce$Day_Exposure)

# Reorder the levels
sce$Drug <- factor(sce$Drug, levels = c("TMZ", "DMSO"))
print(sce$Drug)
```

The dose is related to the number of genes detected

```{r zm_rna_input, fig.width=12}
plt1 <- plotPCA(sce, colour_by="Drug") + 
  theme_Publication()

plt2 <- plotPCA(sce, colour_by="Dose_M", shape_by="Day_Exposure", point_size=2.5) +
  theme_Publication()

plt3 <- plotPCA(sce, colour_by="Day_Exposure", shape_by="Drug", point_size=2.5) + 
  theme_Publication()

plt4 <- plotPCA(sce, colour_by="detected") + 
  theme_Publication()
```

View timepoint and Drug. Set the Drug as a a letter.

```{r PCA_zm_timepoint}
pca_tb <- as_tibble(reducedDim(sce))
pca_tb$Day_Exposure <- sce$Day_Exposure
pca_tb$Drug <- substr(x = sce$Drug, start = 1, stop = 1)

ggplot(pca_tb, aes(x=PC1, y=PC2, colour=Day_Exposure, label=Drug)) +
  geom_text(size=4.5) +
  xlab("PC1 (35%)") + ylab("PC2 (5%)") +
  scale_colour_brewer(palette = "Dark2", name = "Days") +
  theme_Publication()
```

View library size and Drug. Set the Drug as a a shape.

Some relationship with library size but I don't think it explains everything

```{r PCA_zm_lib_size}
pca_tb <- as_tibble(reducedDim(sce))
pca_tb$sum <- log(sce$sum+1)
pca_tb$Drug <- sce$Drug

ggplot(pca_tb, aes(x=PC1, y=PC2, colour=sum, shape=Drug)) +
  geom_point(size=3) +
  xlab("PC1 (35%)") + ylab("PC2 (5%)") +
  theme_Publication()
```

Conventional visualisation

```{r PCA_basic_timepoint}
plotPCA(sce, colour_by="Day_Exposure", shape_by="Drug", point_size=2.5) + 
  theme_Publication()
```

View dose and timepoint

```{r PCA_zm_dose}
plt2
```

## Save subset SCE

```{r}
saveRDS(sce, here::here(
    "data/TIRE_brain_human/SCEs/", "brainCancer_subset.sce.rds"
   ))
```

### Session info
```{r}
sessionInfo()
```
